+++ src/lib_proxy/light_core.ml
@@ -118,7 +118,7 @@ let get_core (module Light_proto : Light_proto.PROTO_RPCS)
       | ((uri, rpc_context) as hd_endpoint) :: tl_remaining_endpoints -> (
           let* proof_opt =
             Light_proto.merkle_tree
-              Proxy.{rpc_context; chain; block; mode = Client}
+              Proxy.{rpc_context; chain; block}
               key
               leaf_kind
           in